Move rpc client to api package
