+++ pkg/protocol/rpcserver.go
@@ -9,6 +9,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog/log"
 
+	"github.com/chronomq/chronomq/api/rpc/chronomq"
 	"github.com/chronomq/chronomq/internal/monitor"
 	"github.com/chronomq/chronomq/pkg/hub"
 	"github.com/chronomq/chronomq/pkg/job"
@@ -23,20 +24,13 @@ type RPCServer struct {
 	hub *hub.Hub
 }
 
-// RPCJob is a light wrapper struct representing job data on the wire without extra metadata that is stored internally
-type RPCJob struct {
-	Body  []byte
-	ID    string
-	Delay time.Duration
-}
-
 func newRPCServer(hub *hub.Hub) *RPCServer {
 	memMonitor = monitor.GetMemMonitor()
 	return &RPCServer{hub: hub}
 }
 
 // PutWithID accepts a new job and stores it in a Hub, reply is ignored
-func (r *RPCServer) PutWithID(rpcJob RPCJob, id *string) error {
+func (r *RPCServer) PutWithID(rpcJob chronomq.RPCJob, id *string) error {
 	memMonitor.Fence()
 
 	var j *job.Job
@@ -64,7 +58,7 @@ func (r *RPCServer) Cancel(id string, ignoredReply *int8) error {
 // Next sets the reply (job) to a valid job if a job is ready to be triggered
 // If not job is ready yet, this call will wait (block) for the given duration and keep searching
 // for ready jobs. If no job is ready by the end of the timeout, ErrTimeout is returned
-func (r *RPCServer) Next(timeout time.Duration, job *RPCJob) error {
+func (r *RPCServer) Next(timeout time.Duration, job *chronomq.RPCJob) error {
 	// try once
 	if j := r.hub.NextLocked(); j != nil {
 		defer memMonitor.Decrement(j)
@@ -107,7 +101,7 @@ func (r *RPCServer) Ping(ignore int8, pong *string) error {
 }
 
 // InspectN returns n jobs without removing them for ad-hoc inspection
-func (r *RPCServer) InspectN(n int, rpcJobs *[]*RPCJob) error {
+func (r *RPCServer) InspectN(n int, rpcJobs *[]*chronomq.RPCJob) error {
 	if n == 0 {
 		return nil
 	}
@@ -115,7 +109,7 @@ func (r *RPCServer) InspectN(n int, rpcJobs *[]*RPCJob) error {
 	jobs := r.hub.GetNJobs(n)
 
 	for j := range jobs {
-		rpcJob := &RPCJob{
+		rpcJob := &chronomq.RPCJob{
 			Body:  j.Body(),
 			ID:    j.ID(),
 			Delay: j.TriggerAt().Sub(time.Now()),
