+++ pkg/protocol/yaad_rpc_integration_test.go
@@ -8,6 +8,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
+	"github.com/chronomq/chronomq/api/rpc/chronomq"
 	"github.com/chronomq/chronomq/pkg/hub"
 	"github.com/chronomq/chronomq/pkg/persistence"
 	"github.com/chronomq/chronomq/pkg/protocol"
@@ -16,7 +17,7 @@ import (
 var _ = Describe("Test rpc protocol:", func() {
 	defer GinkgoRecover()
 	var port = 9001
-	var client *protocol.RPCClient
+	var client *chronomq.RPCClient
 
 	var srv io.Closer
 	var h *hub.Hub
@@ -35,7 +36,7 @@ var _ = Describe("Test rpc protocol:", func() {
 		Expect(err).NotTo(HaveOccurred())
 		port++
 
-		client = &protocol.RPCClient{}
+		client = &chronomq.RPCClient{}
 
 		// This ensures all contexts get a running server
 		Eventually(func() error {
@@ -93,7 +94,7 @@ var _ = Describe("Test rpc protocol:", func() {
 		ExpectNoErr(err)
 
 		// We can inspect without consuming too
-		rpcJobs := []*protocol.RPCJob{}
+		rpcJobs := []*chronomq.RPCJob{}
 		err = client.InspectN(2, &rpcJobs)
 		Expect(err).To(BeNil())
 		Expect(len(rpcJobs)).To(Equal(1))
@@ -121,7 +122,7 @@ var _ = Describe("Test rpc protocol:", func() {
 
 		// InspectN < n
 		inspectN := 5
-		rpcJobs := []*protocol.RPCJob{}
+		rpcJobs := []*chronomq.RPCJob{}
 		err := client.InspectN(inspectN, &rpcJobs)
 		Expect(err).To(BeNil())
 		Expect(len(rpcJobs)).To(Equal(inspectN))
@@ -131,7 +132,7 @@ var _ = Describe("Test rpc protocol:", func() {
 		}
 		// InspectN == n
 		inspectN = n
-		rpcJobs = []*protocol.RPCJob{}
+		rpcJobs = []*chronomq.RPCJob{}
 		err = client.InspectN(inspectN, &rpcJobs)
 		Expect(err).To(BeNil())
 		Expect(len(rpcJobs)).To(Equal(inspectN))
@@ -142,7 +143,7 @@ var _ = Describe("Test rpc protocol:", func() {
 
 		// InspectN > n
 		inspectN = n + 3
-		rpcJobs = []*protocol.RPCJob{}
+		rpcJobs = []*chronomq.RPCJob{}
 		err = client.InspectN(inspectN, &rpcJobs)
 		Expect(err).To(BeNil())
 		Expect(len(rpcJobs)).To(Equal(n))
